+++ src/server-ts/src/index.ts
@@ -1,10 +1,19 @@
-import http from 'http';
-import cors from 'cors';
-import express from 'express';
-
-import InitWebSocket from './core/ws.ts';
+import http from 'http'
+import cors from 'cors'
+import express from 'express'
+// 全局路径
+import {PATHS, __dirname, __filename} from './utils/paths.ts'
+// ws
+import InitWebSocket from './core/ws.ts'
 // routers
-import router from './router/indexRouter.ts';
+import router from './router/indexRouter.ts'
+// config
+import config from './utils/config.ts'
+
+console.log('通过 \'/utils/paths.ts\' 配置的全局路径列表:')
+console.log('PATHS:',PATHS)
+console.log('__dirname:',__dirname)
+console.log('__filename:', __filename)
 
 const app = express();
 const server = http.createServer(app);
@@ -15,8 +24,8 @@ app.use(cors());
 app.use(express.json());
 app.use("/", router);
 
-const ip = "localhost";
-const port = 9119;
+const ip = config.ip ?? "localhost";
+const port = config.port ?? 9119;
 
 server.listen(port, ip, () => {
   console.log(`Server is running at http://${ip}:${port}`);